+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ledgerhq/live-cli
 
+## 24.45.2
+
+### Patch Changes
+
+- Updated dependencies [[`91b83cf`](https://github.com/LedgerHQ/ledger-live/commit/91b83cf8af507a92d2ff535af4e62264a7c142a1)]:
+  - @ledgerhq/live-common@36.3.1
+
+## 24.45.2-hotfix.1
+
+### Patch Changes
+
+- Updated dependencies [[`91b83cf`](https://github.com/LedgerHQ/ledger-live/commit/91b83cf8af507a92d2ff535af4e62264a7c142a1)]:
+  - @ledgerhq/live-common@36.3.1-hotfix.1
+
+## 24.45.2-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/live-common@36.3.1-hotfix.0
+
 ## 24.45.1
 
 ### Patch Changes