+++ apps/ledger-live-mobile/CHANGELOG.md
@@ -1,5 +1,17 @@
 # live-mobile
 
+## 4.10.1
+
+### Patch Changes
+
+- [#19602](https://github.com/LedgerHQ/ledger-live/pull/19602) [`5528d15`](https://github.com/LedgerHQ/ledger-live/commit/5528d15573b9ac1c54ce590b1b26b5cdfdf8eefa) Thanks [@LucasWerey](https://github.com/LucasWerey)! - Fix asset aggreg for L2
+
+## 4.10.1-hotfix.0
+
+### Patch Changes
+
+- [#19602](https://github.com/LedgerHQ/ledger-live/pull/19602) [`5528d15`](https://github.com/LedgerHQ/ledger-live/commit/5528d15573b9ac1c54ce590b1b26b5cdfdf8eefa) Thanks [@LucasWerey](https://github.com/LucasWerey)! - Fix asset aggreg for L2
+
 ## 4.10.0
 
 ### Minor Changes