+++ libs/test-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ledgerhq/test-utils
 
+## 0.4.21
+
+### Patch Changes
+
+- Updated dependencies [[`91b83cf`](https://github.com/LedgerHQ/ledger-live/commit/91b83cf8af507a92d2ff535af4e62264a7c142a1)]:
+  - @ledgerhq/live-common@36.3.1
+
+## 0.4.21-hotfix.1
+
+### Patch Changes
+
+- Updated dependencies [[`91b83cf`](https://github.com/LedgerHQ/ledger-live/commit/91b83cf8af507a92d2ff535af4e62264a7c142a1)]:
+  - @ledgerhq/live-common@36.3.1-hotfix.1
+
+## 0.4.21-hotfix.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/live-common@36.3.1-hotfix.0
+
 ## 0.4.20
 
 ### Patch Changes